# apps/api/api-documentation/swagger.yaml
openapi: 3.0.0
info:
  title: Warehouse API
  version: 1.0.0
  description: API for warehouse management. Only /auth/login endpoint is public. All other endpoints require authentication.
servers:
  - url: http://localhost:3000/api
paths:
...
  /goods/{id}:
    get:
      summary: Get a good by ID
      tags:
        - Goods
      description: Retrieves a good by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: withAmount
          in: query
          description: Include the amount of goods in the response
          schema:
            type: boolean
          required: true
      responses:
        '200':
          description: Good fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoodReadResponse'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '404':
          description: Good not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
    delete:
      summary: Delete a good by ID
      tags:
        - Goods
      description: Deletes a good by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Good deleted successfully
        '404':
          description: Good not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
    post:
      summary: Buy a good by ID
      tags:
        - Goods
      description: Buys a specified amount of a good by its ID.
      operationId: buyGood
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyGoodRequest'
      responses:
        '200':
          description: Good bought successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyGoodResponse'
        '404':
          description: Good not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalError'
    put:
      summary: Sell a good by ID
      tags:
        - Goods
      description: Sells a specified amount of a good by its ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellGoodResponse'
      responses:
        '200':
          description: Good sold successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: integer
                  message:
                    type: string
        '404':
          description: Good not found
...
components:
  schemas:
...
    Good:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        unit:
            $ref: '#/components/schemas/Unit'
        categoryName:
          type: string
      required:
        - id
        - name
        - description
        - unit
        - categoryName
...
    GoodWithAmount:
      allOf:
        - $ref: '#/components/schemas/Good'
        - type: object
      properties:
        amount:
          type: number
        value:
          type: number
    GoodReadResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/GoodWithAmount'
        message:
          type: string
      required:
        - data
        - message
...
    SellGoodResponse:
      type: object
      properties:
        data:
          type: number
        message:
          type: string
      required:
        - data
        - message
    BuyGoodRequest:
      type: object
      properties:
        userId:
          type: number
        amount:
          type: number
        pricePerUnit:
          type: number
      required:
        - userId
        - amount
        - pricePerUnit
    BuyGoodResponse:
      type: object
      properties:
        data:
          type: number
        message:
          type: string
      required:
        - data
        - message
    Unit:
      type: string
      enum:
        - pcs
        - kg
...
    Error:
      type: object
      properties:
        name:
          type: string
        message:
          type: string
        cause:
          type: string
          nullable: true
    ValidationError:
      $ref: '#/components/schemas/Error'
    NotFoundError:
      $ref: '#/components/schemas/Error'
    InternalError:
      $ref: '#/components/schemas/Error'
...